COSETO/requirements.txt
torch 
transformers 
pandas 
python-dotenv 
sqlalchemy
protobuf
tiktoken
SentencePiece
accelerate
